requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.0.6
django-crispy-forms==1.9.1
djangorestframework==3.12.2
future==0.18.2
futures==3.1.1
goslate==1.5.1
idna==2.10
MouseInfo==0.1.3
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.2.0
psycopg2==2.8.6
PyAutoGUI==0.9.50
PyDictionary==2.0.1
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.0
PyTweening==1.0.3
pytz==2020.1
requests==2.24.0
soupsieve==2.0.1
sqlparse==0.3.1
tkintertable==1.3.2
urllib3==1.25.10
virtualenv==16.7.5
